refactor(Gemini): 클래스 책임 분리
- GeminiErrorHandler: Gemini 응답이 4xx 또는 5xx로 왔을 때 예외로 처리하기 위한 Handler.
- GeminiResponseValidator: Gemini의 응답이 유효한 값인지 확인하기 위한 Validator.
